requirements.txt
pytest == 7.3.2
allure-pytest == 2.13.2
selene == 2.0.0b17